requirements.txt
asgiref==3.9.1
Django==5.2.4
django-allauth==65.11.0
django-filter==25.1
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
psycopg2-binary==2.9.10
Pygments==2.19.2
python-dotenv==1.2.1
PyYAML==6.0.2
sqlparse==0.5.3
uritemplate==4.2.0
whitenoise==6.11.0
